LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-image-href.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><image href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=a9fb66e5-f811-42ba-a48c-3fd9156175dd&amp;encodingcheck=&Gbreve;"></image></svg>


FAIL Speculative parsing, document.write(): svg-image-href Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
